fix: preserve ProjectTree size budget
Problem: the session catalog race fix increased ProjectTree.tsx beyond its repository lint budget.

Root cause: the owner-state guards were expanded as new lines inside a file whose existing overage has a zero-growth baseline.

Fix: move the pure guard into the existing catalog presentation module and keep the ProjectTree wiring at its prior line count without changing behavior or updating the lint baseline.

Verification: go run ./tools/repolint; focused catalog ownership and presentation tests; frontend typecheck, test typecheck, and hooks lint.
